Break up persistence package
